fastapi/fastapi/requirements.txt
fastapi==0.103.2
python-dotenv==1.0.0
uvicorn==0.23.2
asyncpg==0.28.0
sqlmodel==0.0.8
alembic==1.12.0
python-jose==3.3.0
python-multipart==0.0.6
